% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotALL.R
\name{get_bar}
\alias{get_bar}
\title{create bar plot}
\usage{
get_bar(x_val, name)
}
\arguments{
\item{x_val}{A vector of factor/integer}

\item{name}{the label of x_val}
}
\value{
a bar plot
}
\description{
create a bar plot for x, with a given lable on x axis
}
\examples{
get_bar(c("a","b","a"),"type")
}
