% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotALL.R
\name{get_hist}
\alias{get_hist}
\title{create histogram}
\usage{
get_hist(x_val, name)
}
\arguments{
\item{x_val}{A vector of doubles}

\item{name}{the label of x_val}
}
\value{
a histogram
}
\description{
create histogram for a vector
}
\examples{
get_hist(rnorm(10),"num")
}
