% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotALL.R
\name{get_scatter}
\alias{get_scatter}
\title{create scatter plot}
\usage{
get_scatter(x_val, y_val, name)
}
\arguments{
\item{x_val}{A vector of doubles}

\item{y_val}{A vector of doubles, must has the same size as x_val}

\item{name}{the label of x_val}
}
\value{
a scatter plot
}
\description{
create scatter plot where y vs x, where y is the price
the linear regression line will also be generated
}
\examples{
get_scatter(rnorm(10),rnorm(10),"num")
}
