% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_tr_tst}
\alias{get_tr_tst}
\title{Take the full dataset into traning or testing set}
\usage{
get_tr_tst(automobile, set = "at")
}
\arguments{
\item{automobile}{the automobile dataframe}

\item{set}{decide the type of dataset.
If "at", all columns except ID;
if "basic", all columns;
if "sub", columns except ID or categorical variables with more than 2 levels}
}
\value{
a list of training dataframes or testing dataframes with or without id
}
\description{
Take the full dataset into traning or testing set
}
\examples{
#get_tr_tst(automobile)
}
