% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_trm_tsm}
\alias{get_trm_tsm}
\title{Get training and testing matrix x,y for lasso and ridge models}
\usage{
get_trm_tsm(training_df_sub, testing_df_sub, set = "training")
}
\arguments{
\item{training_df_sub}{the training set without ID or categorical variables with more than 2 levels}

\item{testing_df_sub}{the concise testing set without ID or categorical variables with more than 2 levels}

\item{set}{whether the desired output is training or testing matrices}
}
\value{
a list of matrices (x and y) for training or testing
}
\description{
Get training and testing matrix x,y for lasso and ridge models
}
\examples{
# get_trm_tsm(training_df_sub, testing_df_sub, "testing")
}
