% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getR2.R
\name{getR2}
\alias{getR2}
\title{Return the n most important variable and their R^2}
\usage{
getR2(dat, n)
}
\arguments{
\item{dat}{A data frame, the response variable must be in the last column}

\item{n}{The numbers of most important variables that will be returned}
}
\value{
A summary table of R^2 and the corresponding variable names
}
\description{
After fitting all explanatory variables respectively in linear model, take the n variables
that have the highest R^2
}
\details{
Create a summary table of the predictor variables with the top n highest R^2 values
}
\examples{
getR2(mtcars, 10)
}
