% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotALL.R
\name{plotAll}
\alias{plotAll}
\title{Show all the plots of the selected predictor variables}
\usage{
plotAll(df, nms)
}
\arguments{
\item{df}{A data frame, where the response variable is in the last column}

\item{nms}{The names of explanatory variables as x-axis}
}
\value{
a list of plots
}
\description{
For the numerical variables, we created both a histogram to see the distribution,
and a scatter plot to see the relationship between the variable and the car price.
}
\details{
For the categorical variables,
we created a bar graph to compare the count of each category in a variable.
}
\examples{
plotAll(mtcars, c("mpg", "cyl"))
}
