% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveVar.R
\name{saveVar}
\alias{saveVar}
\title{Save the variable into file as .rds}
\usage{
saveVar(var, name, out_dir)
}
\arguments{
\item{var}{the variable to be saved}

\item{name}{the expected name of variable, should followed by .rds}

\item{out_dir}{the location of saved variable}
}
\value{
a message of saved variable information
}
\description{
Save the variable into file as .rds
}
\examples{
#saveVar(10,"num.rds","saved")
}
